Shape Keys not moving any meshes with shape keys

+++ source/ncengine/graphics2/diligent/resource/EnvironmentBufferResource.h
@@ -33,6 +33,6 @@ class EnvironmentBufferResource
     private:
         DynamicUniformBuffer m_buffer;
         Diligent::IShaderResourceVariable* m_variable;
-        double m_elapsedSeconds;
+        float m_elapsedSeconds;
 };
 } // namespace nc::graphics
